+++ .ci/travis/run.sh
@@ -24,10 +24,10 @@ else
     python psutil/tests/runner.py
 fi
 
-if [ "$PYVER" == "2.7" ] || [ "$PYVER" == "3.5" ]; then
-    # run mem leaks test
+# Run memory leak tests and linter only on Linux and latest major Python
+# versions.
+if [ "$PYVER" == "2.7" ] || [ "$PYVER" == "3.6" ]; then
     python psutil/tests/test_memory_leaks.py
-    # run linter (on Linux only)
     if [[ "$(uname -s)" != 'Darwin' ]]; then
         python -m flake8
     fi